dep_prev {knitr}	R Documentation

Make later chunks depend on previous chunks

Description

This function can be used to build dependencies among chunks so that all later chunks depend on previous chunks, i.e. whenever the cache of a previous chunk is updated, the cache of all its later chunks will be updated.

Usage

dep_prev()

Value

NULL; the internal dependency structure is updated as a side effect.

References

https://yihui.org/knitr/demo/cache/

See Also

dep_auto

[Package knitr version 1.44 Index]