knit_global {knitr}	R Documentation

The global environment in which code chunks are evaluated

Description

This function makes the environment of a code chunk accessible inside a chunk.

Usage

knit_global()

Details

It returns the envir argument of knit, e.g. if we call knit() in the global environment, knit_global() returns R's global environment by default. You can call functions like ls() on this environment.

[Package knitr version 1.44 Index]