knit_hooks {knitr}	R Documentation

Hooks for R code chunks, inline R code and output

Description

A hook is a function of a pre-defined form (arguments) that takes values of arguments and returns desired output. The object knit_hooks is used to access or set hooks in this package.

Usage

knit_hooks

Format

An object of class list of length 7.

References

Usage: https://yihui.org/knitr/objects/

Components in knit_hooks: https://yihui.org/knitr/hooks/

Examples

knit_hooks$get("source")
knit_hooks$get("inline")

[Package knitr version 1.44 Index]