opts_chunk {knitr}	R Documentation

Default and current chunk options

Description

Options for R code chunks. When running R code, the object opts_chunk (default options) is not modified by chunk headers (local chunk options are merged with default options), whereas opts_current (current options) changes with different chunk headers and it always reflects the options for the current chunk.

Usage

opts_chunk

opts_current

Format

An object of class list of length 7.

An object of class list of length 7.

Details

Normally we set up the global options once in the first code chunk in a document using opts_chunk$set(), so that all latter chunks will use these options. Note the global options set in one chunk will not affect the options in this chunk itself, and that is why we often need to set global options in a separate chunk.

See str(knitr::opts_chunk$get()) for a list of default chunk options.

Note

opts_current should be treated as read-only and you are supposed to only query its values via opts_current$get(). Calling opts_current$set() will throw an error.

References

Usage: https://yihui.org/knitr/objects/

A list of available options: https://yihui.org/knitr/options/#chunk-options

Examples

opts_chunk$get("prompt")
opts_chunk$get("fig.keep")

[Package knitr version 1.44 Index]