opts_hooks {knitr}	R Documentation

Hooks for code chunk options

Description

Like knit_hooks, this object can be used to set hook functions to manipulate chunk options.

Usage

opts_hooks

Format

An object of class list of length 7.

Details

For every code chunk, if the chunk option named, say, FOO, is not NULL, and a hook function with the same name has been set via opts_hooks$set(FOO = function(options) { options }) (you can manipuate the options argument in the function and return it), the hook function will be called to update the chunk options.

References

https://yihui.org/knitr/hooks/

Examples

# make sure the figure width is no smaller than fig.height
opts_hooks$set(fig.width = function(options) {
    if (options$fig.width < options$fig.height) {
        options$fig.width = options$fig.height
    }
    options
})
# remove all hooks
opts_hooks$restore()

[Package knitr version 1.44 Index]