raw_block {knitr}	R Documentation

Mark character strings as raw blocks in R Markdown

Description

Wraps content in a raw attribute block, which protects it from being escaped by Pandoc. See https://pandoc.org/MANUAL.html#generic-raw-attribute. Functions raw_latex() and raw_html() are shorthands of raw_block(x, 'latex') and raw_block(x, 'html'), respectively.

Usage

raw_block(x, type = "latex", ...)

raw_latex(x, ...)

raw_html(x, ...)

Arguments

x
The character vector to be protected.

type
The type of raw blocks (i.e., the Pandoc output format). If you are not sure about the Pandoc output format of your document, insert a code chunk knitr:::pandoc_to() and see what it returns after the document is compiled.

...
Arguments to be passed to asis_output().

Examples

knitr::raw_latex("\\emph{some text}")

[Package knitr version 1.44 Index]