read_rforge {knitr}	R Documentation

Read source code from R-Forge

Description

This function reads source code from the SVN repositories on R-Forge.

Usage

read_rforge(path, project, extra = "")

Arguments

path
Relative path to the source script on R-Forge.

project
Name of the R-Forge project.

extra
Extra parameters to be passed to the URL (e.g. extra = '&revision=48' to check out the source of revision 48).

Value

A character vector of the source code.

Author(s)

Yihui Xie and Peter Ruckdeschel

Examples

library(knitr)
## No test: 
# relies on r-forge.r-project.org being accessible
read_rforge("rgl/R/axes.R", project = "rgl")

##   [1] "# This internal function returns a list with the following components:"                                                 
##   [2] "# xlim, ylim, zlim:  the bounding box expanded so no coordinate has zero or negative extent"                            
##   [3] "# strut:  a boolean indicating whether an expansion was done above"                                                     
##   [4] "# x, y, z:  the box above expanded by a factor of expand"                                                               
##   [5] ""                                                                                                                       
##   [6] ".getRanges <- function(expand = 1.03, ranges=par3d('bbox')) {"                                                          
##   [7] "    ranges <- list(xlim=ranges[1:2], ylim=ranges[3:4], zlim=ranges[5:6])"                                               
##   [8] ""                                                                                                                       
##   [9] "    strut <- FALSE"                                                                                                     
##  [10] "    "                                                                                                                   
##  [11] "    ranges <- lapply(ranges, function(r) {"                                                                             
##  [12] "                       d <- diff(r)"                                                                                    
##  [13] "                       if (d > 0) return(r)"                                                                            
##  [14] "                       strut <<- TRUE"                                                                                  
##  [15] "                       if (d < 0) return(c(0,1))"                                                                       
##  [16] "                       else if (r[1] == 0) return(c(-1, 1))"                                                            
##  [17] "                       else return(r[1] + 0.4*abs(r[1])*c(-1,1))"                                                       
##  [18] "                     })"                                                                                                
##  [19] "                     "                                                                                                  
##  [20] "    ranges$strut <- strut"                                                                                              
##  [21] "    "                                                                                                                   
##  [22] "    ranges$x <- (ranges$xlim - mean(ranges$xlim))*expand + mean(ranges$xlim)"                                           
##  [23] "    ranges$y <- (ranges$ylim - mean(ranges$ylim))*expand + mean(ranges$ylim)"                                           
##  [24] "    ranges$z <- (ranges$zlim - mean(ranges$zlim))*expand + mean(ranges$zlim)"                                           
##  [25] "   "                                                                                                                    
##  [26] "    ranges"                                                                                                             
##  [27] "}"                                                                                                                      
##  [28] ""                                                                                                                       
##  [29] "axis3d <- function(edge, at = NULL, labels = TRUE, tick = TRUE, line = 0,"                                              
##  [30] "    pos = NULL, nticks = 5, ...) {"                                                                                     
##  [31] "        save <- par3d(skipRedraw = TRUE, ignoreExtent = TRUE)"                                                          
##  [32] "        on.exit(par3d(save))"                                                                                           
##  [33] "        "                                                                                                               
##  [34] "\tranges <- .getRanges()"                                                                                               
##  [35] ""                                                                                                                       
##  [36] "\tedge <- c(strsplit(edge, '')[[1]], '-', '-')[1:3]"                                                                    
##  [37] "\tcoord <- match(toupper(edge[1]), c('X', 'Y', 'Z')) "                                                                  
##  [38] "\t"                                                                                                                     
##  [39] "\t# Put the sign in the appropriate entry of edge"                                                                      
##  [40] "\tif (coord == 2) edge[1] <- edge[2]"                                                                                   
##  [41] "\telse if (coord == 3) edge[1:2] <- edge[2:3]"                                                                          
##  [42] "\t"                                                                                                                     
##  [43] "        range <- ranges[[coord]]"                                                                                       
##  [44] ""                                                                                                                       
##  [45] "        if (is.null(at)) {"                                                                                             
##  [46] "                at <- pretty(range, nticks)"                                                                            
##  [47] "                at <- at[at >= range[1] & at <= range[2]]"                                                              
##  [48] "        }"                                                                                                              
##  [49] ""                                                                                                                       
##  [50] "        if (is.logical(labels)) {"                                                                                      
##  [51] "                if (labels) labels <- format(at)"                                                                       
##  [52] "                else labels <- NA"                                                                                      
##  [53] "        }"                                                                                                              
##  [54] ""                                                                                                                       
##  [55] "\tmpos <- matrix(NA,3,length(at))"                                                                                      
##  [56] "\tif (edge[1] == '+') mpos[1,] <- ranges$x[2]"                                                                          
##  [57] "\telse mpos[1,] <- ranges$x[1]"                                                                                         
##  [58] "\tif (edge[2] == '+') mpos[2,] <- ranges$y[2]"                                                                          
##  [59] "\telse mpos[2,] <- ranges$y[1]"                                                                                         
##  [60] "\tif (edge[3] == '+') mpos[3,] <- ranges$z[2]"                                                                          
##  [61] "\telse mpos[3,] <- ranges$z[1]"                                                                                         
##  [62] "                "                                                                                                       
##  [63] "        ticksize <- 0.05*(mpos[,1]-c(mean(ranges$x),mean(ranges$y),mean(ranges$z)))"                                    
##  [64] "        ticksize[coord] <- 0"                                                                                           
##  [65] "               "                                                                                                        
##  [66] "        if (!is.null(pos)) mpos <- matrix(pos,3,length(at))"                                                            
##  [67] "        mpos[coord,] <- at"                                                                                             
##  [68] ""                                                                                                                       
##  [69] "        x <- c(mpos[1,1],mpos[1,length(at)])"                                                                           
##  [70] "        y <- c(mpos[2,1],mpos[2,length(at)])"                                                                           
##  [71] "        z <- c(mpos[3,1],mpos[3,length(at)])"                                                                           
##  [72] "        if (tick) {"                                                                                                    
##  [73] "                x <- c(x,as.double(rbind(mpos[1,],mpos[1,]+ticksize[1])))"                                              
##  [74] "                y <- c(y,as.double(rbind(mpos[2,],mpos[2,]+ticksize[2])))"                                              
##  [75] "                z <- c(z,as.double(rbind(mpos[3,],mpos[3,]+ticksize[3])))"                                              
##  [76] "        }"                                                                                                              
##  [77] "        result <- c(ticks=segments3d(x,y,z,...))"                                                                       
##  [78] ""                                                                                                                       
##  [79] "        if (!all(is.na(labels)))"                                                                                       
##  [80] "                result <- c(result, labels=text3d(mpos[1,]+3*ticksize[1],"                                              
##  [81] "                       mpos[2,]+3*ticksize[2],"                                                                         
##  [82] "                       mpos[3,]+3*ticksize[3],"                                                                         
##  [83] "                       labels, ...))"                                                                                   
##  [84] "\tlowlevel(result)"                                                                                                     
##  [85] "}"                                                                                                                      
##  [86] ""                                                                                                                       
##  [87] "axes3d <- function(edges='bbox', labels=TRUE,"                                                                          
##  [88] "                   tick=TRUE, nticks = 5, box = FALSE, expand = 1.03, ...) {"                                           
##  [89] "    save <- par3d(skipRedraw = TRUE, ignoreExtent = TRUE)"                                                              
##  [90] "    on.exit(par3d(save))"                                                                                               
##  [91] "    if (identical(edges, 'bbox')) {"                                                                                    
##  [92] "    \tbboxargs <- names(formals(rgl.bbox))"                                                                             
##  [93] "    \tbboxargs <- intersect(bboxargs, names(list(...)))"                                                                
##  [94] "    \totherargs <- setdiff(names(list(...)), bboxargs)"                                                                 
##  [95] "    \tbboxargs <- list(...)[bboxargs]"                                                                                  
##  [96] "    \totherargs <- list(...)[otherargs]"                                                                                
##  [97] "        result <- do.call('bbox3d', c(list(draw_front=box, expand=expand), bboxargs, "                                  
##  [98] "                          do.call('.fixMaterialArgs', c(otherargs, list(Params = list(front='lines', back='lines'))))))"
##  [99] "    } else {"                                                                                                           
## [100] "        result <- numeric(0)"                                                                                           
## [101] "    \tfor (e in edges)"                                                                                                 
## [102] "            result <- c(result, axis3d(e,labels=labels,tick=tick,nticks=nticks, ...))"                                  
## [103] "\tnames(result) <- e"                                                                                                   
## [104] "    }\t"                                                                                                                
## [105] "    lowlevel(result)  "                                                                                                 
## [106] "}"                                                                                                                      
## [107] ""                                                                                                                       
## [108] "box3d <- function(...) {"                                                                                               
## [109] "        save <- par3d(ignoreExtent = TRUE)        "                                                                     
## [110] "        on.exit(par3d(save))"                                                                                           
## [111] "        "                                                                                                               
## [112] "        result <- numeric(0)"                                                                                           
## [113] "        "                                                                                                               
## [114] "        ranges <- .getRanges()          "                                                                               
## [115] "        if (ranges$strut) {"                                                                                            
## [116] "            par3d(ignoreExtent = FALSE)"                                                                                
## [117] "            result <- c(result, strut=segments3d(rep(ranges$xlim, c(2,2)),"                                             
## [118] "                       rep(ranges$ylim, c(2,2)),"                                                                       
## [119] "                       rep(ranges$zlim, c(2,2))))"                                                                      
## [120] "            par3d(ignoreExtent = TRUE)"                                                                                 
## [121] "        }"                                                                                                              
## [122] "                       "                                                                                                
## [123] "        x <- c(rep(ranges$x[1],8),rep(ranges$x,4),rep(ranges$x[2],8))"                                                  
## [124] "        y <- c(rep(ranges$y,2),rep(ranges$y,c(2,2)),rep(ranges$y,c(4,4)),"                                              
## [125] "               rep(ranges$y,2),rep(ranges$y,c(2,2)))"                                                                   
## [126] "        z <- c(rep(ranges$z,c(2,2)),rep(ranges$z,2),rep(rep(ranges$z,c(2,2)),2),"                                       
## [127] "               rep(ranges$z,c(2,2)),rep(ranges$z,2))"                                                                   
## [128] "        lowlevel(c(result, lines=segments3d(x,y,z,...)))"                                                               
## [129] "}"                                                                                                                      
## [130] ""                                                                                                                       
## [131] "mtext3d <- function(text, edge, line = 0, at = NULL, pos = NA, ...) {"                                                  
## [132] "        save <- par3d(ignoreExtent = TRUE)"                                                                             
## [133] "        on.exit(par3d(save))"                                                                                           
## [134] ""                                                                                                                       
## [135] "        ranges <- .getRanges()"                                                                                         
## [136] "        "                                                                                                               
## [137] "\tedge <- c(strsplit(edge, '')[[1]], '-', '-')[1:3]"                                                                    
## [138] "\tcoord <- match(toupper(edge[1]), c('X', 'Y', 'Z')) "                                                                  
## [139] "\t"                                                                                                                     
## [140] "\t# Put the sign in the appropriate entry of edge"                                                                      
## [141] "\tif (coord == 2) edge[1] <- edge[2]"                                                                                   
## [142] "\telse if (coord == 3) edge[1:2] <- edge[2:3]"                                                                          
## [143] ""                                                                                                                       
## [144] "        range <- ranges[[coord]]"                                                                                       
## [145] ""                                                                                                                       
## [146] "        if (is.null(at)) at <- mean(range)"                                                                             
## [147] ""                                                                                                                       
## [148] "        newlen <- max(length(text),length(line),length(at))"                                                            
## [149] "        text <- rep(text, len = newlen)"                                                                                
## [150] "        line <- rep(line, len = newlen)"                                                                                
## [151] "        at <- rep(at, len = newlen)"                                                                                    
## [152] ""                                                                                                                       
## [153] "        if (all(is.na(pos))) {"                                                                                         
## [154] "                pos <- matrix(NA,3,length(at))"                                                                         
## [155] "                if (edge[1] == '+') pos[1,] <- ranges$x[2]"                                                             
## [156] "                else pos[1,] <- ranges$x[1]"                                                                            
## [157] "                if (edge[2] == '+') pos[2,] <- ranges$y[2]"                                                             
## [158] "                else pos[2,] <- ranges$y[1]"                                                                            
## [159] "                if (edge[3] == '+') pos[3,] <- ranges$z[2]"                                                             
## [160] "                else pos[3,] <- ranges$z[1]"                                                                            
## [161] "        }"                                                                                                              
## [162] "        else pos <- matrix(pos,3,length(at))"                                                                           
## [163] "        pos[coord,] <- at"                                                                                              
## [164] "        ticksize <- 0.05*(pos[,1]-c(mean(ranges$x),mean(ranges$y),mean(ranges$z)))"                                     
## [165] "        ticksize[coord] <- 0"                                                                                           
## [166] ""                                                                                                                       
## [167] "        text3d(pos[1,]+3*ticksize[1]*line,"                                                                             
## [168] "               pos[2,]+3*ticksize[2]*line,"                                                                             
## [169] "               pos[3,]+3*ticksize[3]*line,"                                                                             
## [170] "               text, ...)"                                                                                              
## [171] "}   "                                                                                                                   
## [172] ""                                                                                                                       
## [173] "title3d <- function(main = NULL, sub = NULL, xlab = NULL, ylab = NULL, "                                                
## [174] "    zlab = NULL, line = NA, ...) {"                                                                                     
## [175] "        save <- par3d(skipRedraw = TRUE, ignoreExtent = TRUE)"                                                          
## [176] "        on.exit(par3d(save))"                                                                                           
## [177] "\tresult <- numeric(0)"                                                                                                 
## [178] "        if (!is.null(main)) {"                                                                                          
## [179] "            aline <- ifelse(is.na(line), 2, line)"                                                                      
## [180] "            result <- c(result, main=mtext3d(main, 'x++', line = aline, ...))"                                          
## [181] "        }"                                                                                                              
## [182] "        if (!is.null(sub)) {"                                                                                           
## [183] "            aline <- ifelse(is.na(line), 3, line)"                                                                      
## [184] "            result <- c(result, sub=mtext3d(sub, 'x', line = aline, ...))"                                              
## [185] "        }"                                                                                                              
## [186] "        if (!is.null(xlab)) {"                                                                                          
## [187] "            aline <- ifelse(is.na(line), 2, line)"                                                                      
## [188] "            result <- c(result, xlab=mtext3d(xlab, 'x', line = aline, ...))"                                            
## [189] "        }"                                                                                                              
## [190] "        if (!is.null(ylab)) {"                                                                                          
## [191] "            aline <- ifelse(is.na(line), 2, line)"                                                                      
## [192] "            result <- c(result, ylab=mtext3d(ylab, 'y', line = aline, ...))"                                            
## [193] "        }"                                                                                                              
## [194] "        if (!is.null(zlab)) {"                                                                                          
## [195] "            aline <- ifelse(is.na(line), 2, line)"                                                                      
## [196] "            result <- c(result, zlab=mtext3d(zlab, 'z', line = aline, ...))"                                            
## [197] "        }                  "                                                                                            
## [198] "        lowlevel(result)"                                                                                               
## [199] "}"

read_rforge("rgl/R/axes.R", project = "rgl", extra = "&revision=519")

##   [1] "# This internal function returns a list with the following components:"                                
##   [2] "# xlim, ylim, zlim:  the bounding box expanded so no coordinate has zero or negative extent"           
##   [3] "# strut:  a boolean indicating whether an expansion was done above"                                    
##   [4] "# x, y, z:  the box above expanded by a factor of expand"                                              
##   [5] ""                                                                                                      
##   [6] ".getRanges <- function(expand = 1.03) {"                                                               
##   [7] "    ranges <- par3d('bbox')"                                                                           
##   [8] "    ranges <- list(xlim=ranges[1:2], ylim=ranges[3:4], zlim=ranges[5:6])"                              
##   [9] ""                                                                                                      
##  [10] "    strut <- FALSE"                                                                                    
##  [11] "    "                                                                                                  
##  [12] "    ranges <- lapply(ranges, function(r) {"                                                            
##  [13] "                       d <- diff(r)"                                                                   
##  [14] "                       if (d > 0) return(r)"                                                           
##  [15] "                       strut <<- TRUE"                                                                 
##  [16] "                       if (d < 0) return(c(0,1))"                                                      
##  [17] "                       else if (r[1] == 0) return(c(-1, 1))"                                           
##  [18] "                       else return(r[1] + 0.4*abs(r[1])*c(-1,1))"                                      
##  [19] "                     })"                                                                               
##  [20] "                     "                                                                                 
##  [21] "    ranges$strut <- strut"                                                                             
##  [22] "    "                                                                                                  
##  [23] "    ranges$x <- (ranges$xlim - mean(ranges$xlim))*expand + mean(ranges$xlim)"                          
##  [24] "    ranges$y <- (ranges$ylim - mean(ranges$ylim))*expand + mean(ranges$ylim)"                          
##  [25] "    ranges$z <- (ranges$zlim - mean(ranges$zlim))*expand + mean(ranges$zlim)"                          
##  [26] "   "                                                                                                   
##  [27] "    ranges"                                                                                            
##  [28] "}"                                                                                                     
##  [29] ""                                                                                                      
##  [30] "axis3d <- function (edge, at = NULL, labels = TRUE, tick = TRUE, line = 0,"                            
##  [31] "    pos = NULL, ...)"                                                                                  
##  [32] "{"                                                                                                     
##  [33] "        save <- par3d(skipRedraw = TRUE, ignoreExtent = TRUE)"                                         
##  [34] "        on.exit(par3d(save))"                                                                          
##  [35] "        "                                                                                              
##  [36] "\tranges <- .getRanges()"                                                                              
##  [37] ""                                                                                                      
##  [38] "\tedge <- c(strsplit(edge, '')[[1]], '-', '-')[1:3]"                                                   
##  [39] "\tcoord <- match(toupper(edge[1]), c('X', 'Y', 'Z')) "                                                 
##  [40] "\t"                                                                                                    
##  [41] "\t# Put the sign in the appropriate entry of edge"                                                     
##  [42] "\tif (coord == 2) edge[1] <- edge[2]"                                                                  
##  [43] "\telse if (coord == 3) edge[1:2] <- edge[2:3]"                                                         
##  [44] "\t"                                                                                                    
##  [45] "        range <- ranges[[coord]]"                                                                      
##  [46] ""                                                                                                      
##  [47] "        if (is.null(at)) {"                                                                            
##  [48] "                at <- pretty(range)"                                                                   
##  [49] "                at <- at[at >= range[1] & at <= range[2]]"                                             
##  [50] "        }"                                                                                             
##  [51] ""                                                                                                      
##  [52] "        if (is.logical(labels)) {"                                                                     
##  [53] "                if (labels) labels <- as.character(at)"                                                
##  [54] "                else labels <- NA"                                                                     
##  [55] "        }"                                                                                             
##  [56] ""                                                                                                      
##  [57] "\tmpos <- matrix(NA,3,length(at))"                                                                     
##  [58] "\tif (edge[1] == '+') mpos[1,] <- ranges$x[2]"                                                         
##  [59] "\telse mpos[1,] <- ranges$x[1]"                                                                        
##  [60] "\tif (edge[2] == '+') mpos[2,] <- ranges$y[2]"                                                         
##  [61] "\telse mpos[2,] <- ranges$y[1]"                                                                        
##  [62] "\tif (edge[3] == '+') mpos[3,] <- ranges$z[2]"                                                         
##  [63] "\telse mpos[3,] <- ranges$z[1]"                                                                        
##  [64] "                "                                                                                      
##  [65] "        ticksize <- 0.05*(mpos[,1]-c(mean(ranges$x),mean(ranges$y),mean(ranges$z)))"                   
##  [66] "        ticksize[coord] <- 0"                                                                          
##  [67] "               "                                                                                       
##  [68] "        if (!is.null(pos)) mpos <- matrix(pos,3,length(at))"                                           
##  [69] "        mpos[coord,] <- at"                                                                            
##  [70] ""                                                                                                      
##  [71] "        x <- c(mpos[1,1],mpos[1,length(at)])"                                                          
##  [72] "        y <- c(mpos[2,1],mpos[2,length(at)])"                                                          
##  [73] "        z <- c(mpos[3,1],mpos[3,length(at)])"                                                          
##  [74] "        if (tick) {"                                                                                   
##  [75] "                x <- c(x,as.double(rbind(mpos[1,],mpos[1,]+ticksize[1])))"                             
##  [76] "                y <- c(y,as.double(rbind(mpos[2,],mpos[2,]+ticksize[2])))"                             
##  [77] "                z <- c(z,as.double(rbind(mpos[3,],mpos[3,]+ticksize[3])))"                             
##  [78] "        }"                                                                                             
##  [79] "        result <- c(ticks=segments3d(x,y,z,...))"                                                      
##  [80] ""                                                                                                      
##  [81] "        if (!all(is.na(labels)))"                                                                      
##  [82] "                result <- c(result, labels=text3d(mpos[1,]+3*ticksize[1],"                             
##  [83] "                       mpos[2,]+3*ticksize[2],"                                                        
##  [84] "                       mpos[3,]+3*ticksize[3],"                                                        
##  [85] "                       labels, ...))"                                                                  
##  [86] "\tinvisible(result)"                                                                                   
##  [87] "}"                                                                                                     
##  [88] ""                                                                                                      
##  [89] "axes3d <- function(edges='bbox', labels=TRUE,"                                                         
##  [90] "                   tick=TRUE, ...)"                                                                    
##  [91] "{"                                                                                                     
##  [92] "    save <- par3d(skipRedraw = TRUE, ignoreExtent = TRUE)"                                             
##  [93] "    on.exit(par3d(save))"                                                                              
##  [94] "    if (identical(edges, 'bbox')) {"                                                                   
##  [95] "        result <- do.call('bbox3d', .fixMaterialArgs(..., Params = list(front='lines', back='lines')))"
##  [96] "    } else {"                                                                                          
##  [97] "        result <- numeric(0)"                                                                          
##  [98] "    \tfor (e in edges)"                                                                                
##  [99] "            result <- c(result, axis3d(e,labels=labels,tick=tick, ...))"                               
## [100] "\tnames(result) <- e"                                                                                  
## [101] "    }\t"                                                                                               
## [102] "    invisible(result)  "                                                                               
## [103] "}"                                                                                                     
## [104] ""                                                                                                      
## [105] "box3d <- function(...)"                                                                                
## [106] "{"                                                                                                     
## [107] "        save <- par3d(ignoreExtent = TRUE)        "                                                    
## [108] "        on.exit(par3d(save))"                                                                          
## [109] "        "                                                                                              
## [110] "        result <- numeric(0)"                                                                          
## [111] "        "                                                                                              
## [112] "        ranges <- .getRanges()          "                                                              
## [113] "        if (ranges$strut) {"                                                                           
## [114] "            par3d(ignoreExtent = FALSE)"                                                               
## [115] "            result <- c(result, strut=segments3d(rep(ranges$xlim, c(2,2)),"                            
## [116] "                       rep(ranges$ylim, c(2,2)),"                                                      
## [117] "                       rep(ranges$zlim, c(2,2))))"                                                     
## [118] "            par3d(ignoreExtent = TRUE)"                                                                
## [119] "        }"                                                                                             
## [120] "                       "                                                                               
## [121] "        x <- c(rep(ranges$x[1],8),rep(ranges$x,4),rep(ranges$x[2],8))"                                 
## [122] "        y <- c(rep(ranges$y,2),rep(ranges$y,c(2,2)),rep(ranges$y,c(4,4)),"                             
## [123] "               rep(ranges$y,2),rep(ranges$y,c(2,2)))"                                                  
## [124] "        z <- c(rep(ranges$z,c(2,2)),rep(ranges$z,2),rep(rep(ranges$z,c(2,2)),2),"                      
## [125] "               rep(ranges$z,c(2,2)),rep(ranges$z,2))"                                                  
## [126] "        invisible(c(result, lines=segments3d(x,y,z,...)))"                                             
## [127] "}"                                                                                                     
## [128] ""                                                                                                      
## [129] "mtext3d <- function(text, edge, line = 0, at = NULL, pos = NA, ...)"                                   
## [130] "{"                                                                                                     
## [131] "        save <- par3d(ignoreExtent = TRUE)"                                                            
## [132] "        on.exit(par3d(save))"                                                                          
## [133] ""                                                                                                      
## [134] "        ranges <- .getRanges()"                                                                        
## [135] "        "                                                                                              
## [136] "\tedge <- c(strsplit(edge, '')[[1]], '-', '-')[1:3]"                                                   
## [137] "\tcoord <- match(toupper(edge[1]), c('X', 'Y', 'Z')) "                                                 
## [138] "\t"                                                                                                    
## [139] "\t# Put the sign in the appropriate entry of edge"                                                     
## [140] "\tif (coord == 2) edge[1] <- edge[2]"                                                                  
## [141] "\telse if (coord == 3) edge[1:2] <- edge[2:3]"                                                         
## [142] ""                                                                                                      
## [143] "        range <- ranges[[coord]]"                                                                      
## [144] ""                                                                                                      
## [145] "        if (is.null(at)) at <- mean(range)"                                                            
## [146] ""                                                                                                      
## [147] "        newlen <- max(length(text),length(line),length(at))"                                           
## [148] "        text <- rep(text, len = newlen)"                                                               
## [149] "        line <- rep(line, len = newlen)"                                                               
## [150] "        at <- rep(at, len = newlen)"                                                                   
## [151] ""                                                                                                      
## [152] "        if (all(is.na(pos))) {"                                                                        
## [153] "                pos <- matrix(NA,3,length(at))"                                                        
## [154] "                if (edge[1] == '+') pos[1,] <- ranges$x[2]"                                            
## [155] "                else pos[1,] <- ranges$x[1]"                                                           
## [156] "                if (edge[2] == '+') pos[2,] <- ranges$y[2]"                                            
## [157] "                else pos[2,] <- ranges$y[1]"                                                           
## [158] "                if (edge[3] == '+') pos[3,] <- ranges$z[2]"                                            
## [159] "                else pos[3,] <- ranges$z[1]"                                                           
## [160] "        }"                                                                                             
## [161] "        else pos <- matrix(pos,3,length(at))"                                                          
## [162] "        pos[coord,] <- at"                                                                             
## [163] "        ticksize <- 0.05*(pos[,1]-c(mean(ranges$x),mean(ranges$y),mean(ranges$z)))"                    
## [164] "        ticksize[coord] <- 0"                                                                          
## [165] ""                                                                                                      
## [166] "        invisible(text3d(pos[1,]+3*ticksize[1]*line,"                                                  
## [167] "               pos[2,]+3*ticksize[2]*line,"                                                            
## [168] "               pos[3,]+3*ticksize[3]*line,"                                                            
## [169] "               text, ...))"                                                                            
## [170] "}   "                                                                                                  
## [171] ""                                                                                                      
## [172] "title3d <- function (main = NULL, sub = NULL, xlab = NULL, ylab = NULL, "                              
## [173] "    zlab = NULL, line = NA, ...) "                                                                     
## [174] "{"                                                                                                     
## [175] "        save <- par3d(skipRedraw = TRUE, ignoreExtent = TRUE)"                                         
## [176] "        on.exit(par3d(save))"                                                                          
## [177] "\tresult <- numeric(0)"                                                                                
## [178] "        if (!is.null(main)) {"                                                                         
## [179] "            aline <- ifelse(is.na(line), 2, line)"                                                     
## [180] "            result <- c(result, main=mtext3d(main, 'x++', line = aline, ...))"                         
## [181] "        }"                                                                                             
## [182] "        if (!is.null(sub)) {"                                                                          
## [183] "            aline <- ifelse(is.na(line), 3, line)"                                                     
## [184] "            result <- c(result, sub=mtext3d(sub, 'x', line = aline, ...))"                             
## [185] "        }"                                                                                             
## [186] "        if (!is.null(xlab)) {"                                                                         
## [187] "            aline <- ifelse(is.na(line), 2, line)"                                                     
## [188] "            result <- c(result, xlab=mtext3d(xlab, 'x', line = aline, ...))"                           
## [189] "        }"                                                                                             
## [190] "        if (!is.null(ylab)) {"                                                                         
## [191] "            aline <- ifelse(is.na(line), 2, line)"                                                     
## [192] "            result <- c(result, ylab=mtext3d(ylab, 'y', line = aline, ...))"                           
## [193] "        }"                                                                                             
## [194] "        if (!is.null(zlab)) {"                                                                         
## [195] "            aline <- ifelse(is.na(line), 2, line)"                                                     
## [196] "            result <- c(result, zlab=mtext3d(zlab, 'z', line = aline, ...))"                           
## [197] "        }                  "                                                                           
## [198] "        invisible(result)"                                                                             
## [199] "}"

## End(No test)

[Package knitr version 1.36 Index]