rst2pdf {knitr}	R Documentation

A wrapper for rst2pdf

Description

Convert reST to PDF using rst2pdf (which converts from rst to PDF using the ReportLab open-source library).

Usage

rst2pdf(input, command = "rst2pdf", options = "")

Arguments

input
The input rst file.

command
Character string giving the path of the rst2pdf program. If the program is not in your PATH, the full path has to be given here.

options
Extra command line options, e.g. '-v'.

Value

An input file ‘*.rst’ will produce ‘*.pdf’ and this output filename is returned if the conversion was successful.

Author(s)

Alex Zvoleff and Yihui Xie

References

https://github.com/rst2pdf/rst2pdf

See Also

knit2pdf

[Package knitr version 1.44 Index]