sew {knitr}	R Documentation

Wrap evaluated results for output

Description

This function is mainly for internal use: it is called on each part of the output of the code chunk (code, messages, text output, and plots, etc.) after all statements in the code chunk have been evaluated, and will sew these pieces of output together into a character vector.

Usage

sew(x, options = list(), ...)

Arguments

x
Output from evaluate::evaluate().

options
A list of chunk options used to control output.

...
Other arguments to pass to methods.

[Package knitr version 1.36 Index]