write_bib {knitr}	R Documentation

Generate BibTeX bibliography databases for R packages

Description

This function uses utils::citation() and utils::toBibtex() to create bib entries for R packages and write them in a file. It can facilitate the auto-generation of bibliography databases for R packages, and it is easy to regenerate all the citations after updating R packages.

Usage

write_bib(
  x = .packages(),
  file = "",
  tweak = TRUE,
  width = NULL,
  prefix = getOption("knitr.bib.prefix", "R-"),
  lib.loc = NULL
)

Arguments

x
Package names. Packages which are not installed are ignored.

file
The (‘.bib’) file to write. By default, or if NULL, output is written to the R console.

tweak
Whether to fix some known problems in the citations, especially non-standard format of author names.

width
Width of lines in bibliography entries. If NULL, lines will not be wrapped.

prefix
Prefix string for keys in BibTeX entries; by default, it is R- unless option('knitr.bib.prefix') has been set to another string.

lib.loc
A vector of path names of R libraries.

Details

For a package, the keyword R-pkgname is used for its bib item, where pkgname is the name of the package. Citation entries specified in the ‘CITATION’ file of the package are also included. The main purpose of this function is to automate the generation of the package citation information because it often changes (e.g. author, year, package version, ...).

Value

A list containing the citations. Citations are also written to the file as a side effect.

Note

Some packages on CRAN do not have standard bib entries, which was once reported by Michael Friendly at https://stat.ethz.ch/pipermail/r-devel/2010-November/058977.html. I find this a real pain, and there are no easy solutions except contacting package authors to modify their DESCRIPTION files. Anyway, the argument tweak has provided ugly hacks to deal with packages which are known to be non-standard in terms of the format of citations; tweak = TRUE is by no means intended to hide or modify the original citation information. It is just due to the loose requirements on package authors for the DESCRIPTION file. On one hand, I apologize if it really mangles the information about certain packages; on the other, I strongly recommend package authors to consider the Authors@R field (see the manual Writing R Extensions) to make it easier for other people to cite R packages. See knitr:::.tweak.bib for details of tweaks. Also note this is subject to future changes since R packages are being updated. If you want to contribute more tweaks, please edit the file ‘inst/misc/tweak_bib.csv’ in the source package.

Author(s)

Yihui Xie and Michael Friendly

Examples

write_bib(c("RGtk2", "gWidgets"), file = "R-GUI-pkgs.bib")

## Warning in write_bib(c("RGtk2", "gWidgets"), file = "R-GUI-pkgs.bib"): package(s) RGtk2, gWidgets not found

unlink("R-GUI-pkgs.bib")

write_bib(c("animation", "rgl", "knitr", "ggplot2"))

## Warning in write_bib(c("animation", "rgl", "knitr", "ggplot2")): package(s) animation, rgl not found

## @Manual{R-ggplot2,
##   title = {ggplot2: Create Elegant Data Visualisations Using the Grammar of Graphics},
##   author = {Hadley Wickham and Winston Chang and Lionel Henry and Thomas Lin Pedersen and Kohske Takahashi and Claus Wilke and Kara Woo and Hiroaki Yutani and Dewey Dunnington},
##   year = {2021},
##   note = {R package version 3.3.5},
##   url = {https://CRAN.R-project.org/package=ggplot2},
## }
## 
## @Manual{R-knitr,
##   title = {knitr: A General-Purpose Package for Dynamic Report Generation in R},
##   author = {Yihui Xie},
##   year = {2021},
##   note = {R package version 1.36},
##   url = {https://yihui.org/knitr/},
## }
## 
## @Book{ggplot22016,
##   author = {Hadley Wickham},
##   title = {ggplot2: Elegant Graphics for Data Analysis},
##   publisher = {Springer-Verlag New York},
##   year = {2016},
##   isbn = {978-3-319-24277-4},
##   url = {https://ggplot2.tidyverse.org},
## }
## 
## @Book{knitr2015,
##   title = {Dynamic Documents with {R} and knitr},
##   author = {Yihui Xie},
##   publisher = {Chapman and Hall/CRC},
##   address = {Boca Raton, Florida},
##   year = {2015},
##   edition = {2nd},
##   note = {ISBN 978-1498716963},
##   url = {https://yihui.org/knitr/},
## }
## 
## @InCollection{knitr2014,
##   booktitle = {Implementing Reproducible Computational Research},
##   editor = {Victoria Stodden and Friedrich Leisch and Roger D. Peng},
##   title = {knitr: A Comprehensive Tool for Reproducible Research in {R}},
##   author = {Yihui Xie},
##   publisher = {Chapman and Hall/CRC},
##   year = {2014},
##   note = {ISBN 978-1466561595},
##   url = {http://www.crcpress.com/product/isbn/9781466561595},
## }

write_bib(c("base", "parallel", "MASS"))  # base and parallel are identical

## @Manual{R-base,
##   title = {R: A Language and Environment for Statistical Computing},
##   author = {{R Core Team}},
##   organization = {R Foundation for Statistical Computing},
##   address = {Vienna, Austria},
##   year = {2021},
##   url = {https://www.R-project.org/},
## }
## 
## @Manual{R-MASS,
##   title = {MASS: Support Functions and Datasets for Venables and Ripley's MASS},
##   author = {Brian Ripley},
##   year = {2021},
##   note = {R package version 7.3-54},
##   url = {http://www.stats.ox.ac.uk/pub/MASS4/},
## }
## 
## @Book{MASS2002,
##   title = {Modern Applied Statistics with S},
##   author = {W. N. Venables and B. D. Ripley},
##   publisher = {Springer},
##   edition = {Fourth},
##   address = {New York},
##   year = {2002},
##   note = {ISBN 0-387-95457-0},
##   url = {https://www.stats.ox.ac.uk/pub/MASS4/},
## }

write_bib("cluster", prefix = "")  # a empty prefix

## @Manual{cluster,
##   title = {cluster: "Finding Groups in Data": Cluster Analysis Extended Rousseeuw et
## al.},
##   author = {Martin Maechler and Peter Rousseeuw and Anja Struyf and Mia Hubert},
##   year = {2021},
##   note = {R package version 2.1.2},
##   url = {https://svn.r-project.org/R-packages/trunk/cluster/},
## }

write_bib("digest", prefix = "R-pkg-")  # a new prefix

## tweaking digest

## @Manual{R-pkg-digest,
##   title = {digest: Create Compact Hash Digests of R Objects},
##   author = {Dirk Eddelbuettel},
##   year = {2021},
##   note = {R package version 0.6.29},
##   url = {https://CRAN.R-project.org/package=digest},
## }

write_bib("digest", tweak = FALSE)  # original version

## @Manual{R-digest,
##   title = {digest: Create Compact Hash Digests of R Objects},
##   author = {Dirk Eddelbuettel with contributions by Antoine Lucas and Jarek Tuszynski and Henrik Bengtsson and Simon Urbanek and Mario Frasca and Bryan Lewis and Murray Stokely and Hannes Muehleisen and Duncan Murdoch and Jim Hester and Wush Wu and Qiang Kou and Thierry Onkelinx and Michel Lang and Viliam Simko and Kurt Hornik and Radford Neal and Kendon Bell and Matthew {de Queljoe} and Ion Suruceanu and Bill Denney and Dirk Schumacher and and Winston Chang.},
##   year = {2021},
##   note = {R package version 0.6.29},
##   url = {https://CRAN.R-project.org/package=digest},
## }

# what tweak=TRUE does
str(knitr:::.tweak.bib)

## List of 43
##  $ ade4         : Named chr "  author = {Stéphane Dray and Anne-Béatrice Dufour and Jean Thioulouse and Thibaut Jombart and Sandrine Pavoine"| __truncated__
##   ..- attr(*, "names")= chr "author"
##  $ akima        : Named chr "  author = {H. Akima and Albrecht Gebhardt and Thomas Petzoldt and Martin Maechler},"
##   ..- attr(*, "names")= chr "author"
##  $ ash          : Named chr "  author = {David W. Scott and Albrecht Gebhardt and Stephen Kaluzny},"
##   ..- attr(*, "names")= chr "author"
##  $ bcpa         : Named chr "  author = {Jose Claudio Faria and Clarice Garcia Borges Demetrio},"
##   ..- attr(*, "names")= chr "author"
##  $ BiplotGUI    : Named chr "  author = {Anthony la Grange and  N. J. le Roux and P.J. Rousseeuw and I. Ruts and J. W. Tukey},"
##   ..- attr(*, "names")= chr "author"
##  $ bitops       : Named chr "  author = {Steve Dutky and Martin Maechler},"
##   ..- attr(*, "names")= chr "author"
##  $ cacheSweave  : Named chr "  author = {Roger D. Peng},"
##   ..- attr(*, "names")= chr "author"
##  $ cat          : Named chr "  author = {Ted Harding and Fernando Tusell and Joseph L. Schafer},"
##   ..- attr(*, "names")= chr "author"
##  $ CircStats    : Named chr "  author = {Ulric Lund and Claudio Agostinelli},"
##   ..- attr(*, "names")= chr "author"
##  $ contrast     : Named chr "  author = {Max Kuhn and Steve Weston and Jed Wing and James Forester},"
##   ..- attr(*, "names")= chr "author"
##  $ date         : Named chr "  author = {Terry Therneau and Thomas Lumley and Kjetil Halvorsen and Kurt Hornik},"
##   ..- attr(*, "names")= chr "author"
##  $ digest       : Named chr "  author = {Dirk Eddelbuettel},"
##   ..- attr(*, "names")= chr "author"
##  $ ElemStatLearn: Named chr "  author = {Kjetil Halvorsen},"
##   ..- attr(*, "names")= chr "author"
##  $ epiR         : Named chr "  author = {Mark Stevenson and Telmo Nunes and Cord Heuer and Jonathon Marshall and Javier Sanchez and Ron Thor"| __truncated__
##   ..- attr(*, "names")= chr "author"
##  $ Fahrmeir     : Named chr "  author = {Kjetil Halvorsen},"
##   ..- attr(*, "names")= chr "author"
##  $ flashClust   : Named chr "  author = {Fionn Murtagh and {R development team} and Peter Langfelder},"
##   ..- attr(*, "names")= chr "author"
##  $ foreach      : Named chr "  author = {{Revolution Analytics} and Steve Weston}},"
##   ..- attr(*, "names")= chr "author"
##  $ fortunes     : Named chr "  author = {Achim Zeileis and the R community},"
##   ..- attr(*, "names")= chr "author"
##  $ gee          : Named chr "  author = {Vincent J Carey and Thomas Lumley and Brian Ripley},"
##   ..- attr(*, "names")= chr "author"
##  $ gmodels      : Named chr "  author = {Gregory R. Warnes andBen Bolker and Thomas Lumley and Randall C Johnson and Randall C. Johnson},"
##   ..- attr(*, "names")= chr "author"
##  $ gWidgets     : Named chr "  author = {John Verzani},"
##   ..- attr(*, "names")= chr "author"
##  $ hexbin       : Named chr "  author = {Dan Carr and Nicholas Lewin-Koh and Martin Maechler},"
##   ..- attr(*, "names")= chr "author"
##  $ Hmisc        : Named chr "  author = {Harrell, Jr., Frank E},"
##   ..- attr(*, "names")= chr "author"
##  $ Hmisc        : Named chr "  author = {Frank E. {Harrell, Jr.}},"
##   ..- attr(*, "names")= chr "author"
##  $ leaps        : Named chr "  author = {Thomas Lumley},"
##   ..- attr(*, "names")= chr "author"
##  $ mapproj      : Named chr "  author = {Doug McIlroy and Ray Brownrigg and Thomas P Minka and Roger Bivand},"
##   ..- attr(*, "names")= chr "author"
##  $ maps         : Named chr "  author = {Ray Brownrigg},"
##   ..- attr(*, "names")= chr "author"
##  $ mathgraph    : Named chr "  author = {Patrick J. Burns and Nick Efthymiou and Claus Dethlefsen},"
##   ..- attr(*, "names")= chr "author"
##  $ oz           : Named chr "  author = {Bill Venables and Kurt Hornik},"
##   ..- attr(*, "names")= chr "author"
##  $ pbivnorm     : Named chr "  author = {Alan Genz and Brenton Kenkel},"
##   ..- attr(*, "names")= chr "author"
##  $ pscl         : Named chr "  author = {Simon Jackman and Alex Tahk and Achim Zeileis and Christina Maimone and Jim Fearon},"
##   ..- attr(*, "names")= chr "author"
##  $ quadprog     : Named chr "  author = {Berwin A. Turlach and Andreas Weingessel},"
##   ..- attr(*, "names")= chr "author"
##  $ R2SWF        : Named chr "  author = {Yixuan Qiu and Yihui Xie and Cameron Bracken},"
##   ..- attr(*, "names")= chr "author"
##  $ R2WinBUGS    : Named chr "  author = {Andrew Gelman and Sibylle Sturtz and Uwe Ligges and Gregor Gorjanc and Jouni Kerman},"
##   ..- attr(*, "names")= chr "author"
##  $ randomForest : Named chr "  author = {Leo Breiman and Adele Cutler and Andy Liaw and Matthew Wiener},"
##   ..- attr(*, "names")= chr "author"
##  $ rgl          : Named chr "  author = {Daniel Adler and Duncan Murdoch},"
##   ..- attr(*, "names")= chr "author"
##  $ RgoogleMaps  : Named chr "  author = {Markus Loecher},"
##   ..- attr(*, "names")= chr "author"
##  $ rms          : Named chr "  author = {Frank E. {Harrell, Jr.}},"
##   ..- attr(*, "names")= chr "author"
##  $ robustbase   : Named chr "  author = {Valentin Todorov and Andreas Ruckstuhl and Matias Salibian-Barrera and Tobias Verbeke and Manuel Ko"| __truncated__
##   ..- attr(*, "names")= chr "author"
##  $ RODBC        : Named chr "  author = {Brian Ripley and Michael Lapsley},"
##   ..- attr(*, "names")= chr "author"
##  $ Sleuth2      : Named chr "  author = {F. L. Ramsey and D. W. Schafer and Jeannie Sifneos and Berwin A. Turlach},"
##   ..- attr(*, "names")= chr "author"
##  $ sm           : Named chr "  author = {Adrian Bowman and Adelchi Azzalini},"
##   ..- attr(*, "names")= chr "author"
##  $ tuneR        : Named chr "  author = {Uwe Ligges},"
##   ..- attr(*, "names")= chr "author"

[Package knitr version 1.36 Index]