current_input {knitr}	R Documentation

Query the current input filename

Description

Returns the name of the input file passed to knit().

Usage

current_input(dir = FALSE)

Arguments

dir
Boolean; whether to prepend the current working directory to the file path, i.e. whether to return an absolute path or a relative path.

Value

A character string, if this function is called inside an input document. Otherwise NULL.

[Package knitr version 1.36 Index]