engine_output {knitr}	R Documentation

An output wrapper for language engine output

Description

If you have designed a language engine, you may call this function in the end to format and return the text output from your engine.

Usage

engine_output(options, code, out, extra = NULL)

Arguments

options
a list of chunk options (usually this is just the object options passed to the engine function; see knit_engines)

code
the source code of the chunk, to which the output hook source is applied, unless the chunk option echo == FALSE

out
the text output from the engine, to which the hook output is applied, unless the chunk option results == 'hide'

extra
any additional text output that you want to include

Value

A character string generated from the source code and output using the appropriate output hooks.

[Package knitr version 1.14 Index]