hook_ffmpeg_html {knitr}	R Documentation

Hooks to create animations in HTML output

Description

hook_ffmpeg_html() uses FFmpeg to convert images to a video; hook_gifski() uses the gifski to convert images to a GIF animation; hook_scianimator() uses the JavaScript library SciAnimator to create animations; hook_r2swf() uses the R2SWF package.

Usage

hook_ffmpeg_html(x, options)

hook_gifski(x, options)

hook_scianimator(x, options)

hook_r2swf(x, options)

Arguments

x
Filename for the plot (a character string).

options
A list of the current chunk options.

Details

These hooks are mainly for the package option animation.fun, e.g. you can set opts_knit$set(animation.fun = hook_scianimator).

[Package knitr version 1.36 Index]