include_url {knitr}	R Documentation

Embed a URL as an HTML iframe or a screenshot in knitr documents

Description

When the output format is HTML, include_url() inserts an iframe in the output; otherwise it takes a screenshot of the URL and insert the image in the output. include_app() takes the URL of a Shiny app and adds ?showcase=0 to it (to disable the showcase mode), then passes the URL to include_url().

Usage

include_url(url, height = "400px")

include_app(url, height = "400px")

Arguments

url
A character vector of URLs.

height
A character vector to specify the height of iframes.

Value

An R object with a special class that knitr recognizes internally to generate the iframes or screenshots.

See Also

include_graphics

[Package knitr version 1.36 Index]