knit_code {knitr}	R Documentation

The code manager to manage code in all chunks

Description

This object provides methods to manage code (as character vectors) in all chunks in knitr source documents. For example, knitr::knit_code$get() returns a named list of all code chunks (the names are chunk labels), and knitr::knit_code$get('foo') returns the character vector of the code in the chunk with the label foo.

Usage

knit_code

Format

An object of class list of length 6.

Note

The methods on this object include the set() method (i.e., you could do something like knitr::knit_code$set(foo = "'my precious new code'")), but we recommend that you do not use this method to modify the content of code chunks, unless you are as creative as Emi Tanaka and know what you are doing.

[Package knitr version 1.36 Index]