knit_params_yaml {knitr}	R Documentation

Extract knit parameters from YAML text

Description

This function reads the YAML front-matter that has already been extracted from a document and returns a list of any parameters declared there.

Usage

knit_params_yaml(yaml, evaluate = TRUE)

Arguments

yaml
Character vector containing the YAML text.

evaluate
If TRUE (the default) expression values embedded within the YAML will be evaluated. If FALSE, parameters defined with an expression will have the parsed but unevaluated expression in their value field.

Value

List of objects of class knit_param that correspond to the parameters declared in the params section of the YAML. See knit_params for a full description of these objects.

See Also

knit_params

[Package knitr version 1.36 Index]