knit_theme {knitr}	R Documentation

Syntax highlighting themes

Description

This object can be used to set or get themes in knitr for syntax highlighting.

Usage

knit_theme

Format

An object of class list of length 2.

Details

We can use knit_theme$set(theme) to set the theme, and knit_theme$get(theme) to get a theme. The theme is a character string for both methods (either the name of the theme, or the path to the CSS file of a theme), and for the set() method, it can also be a list returned by the get() method. See examples below.

Note

The syntax highlighting here only applies to ‘.Rnw’ (LaTeX) and ‘.Rhtml’ (HTML) documents, and it does not work for other types of documents, such as ‘.Rmd’ (R Markdown, which has its own syntax highlighting themes; see https://rmarkdown.rstudio.com).

Author(s)

Ramnath Vaidyanathan and Yihui Xie

References

For a preview of all themes, see https://gist.github.com/yihui/3422133.

Examples

opts_knit$set(out.format = "latex")
knit_theme$set("edit-vim")

knit_theme$get()  # names of all available themes

##  [1] "acid"              "aiseered"          "andes"             "anotherdark"       "autumn"           
##  [6] "baycomb"           "bclear"            "biogoo"            "bipolar"           "blacknblue"       
## [11] "bluegreen"         "breeze"            "bright"            "camo"              "candy"            
## [16] "clarity"           "dante"             "darkblue"          "darkbone"          "darkness"         
## [21] "darkslategray"     "darkspectrum"      "default"           "denim"             "dusk"             
## [26] "earendel"          "easter"            "edit-anjuta"       "edit-eclipse"      "edit-emacs"       
## [31] "edit-flashdevelop" "edit-gedit"        "edit-jedit"        "edit-kwrite"       "edit-matlab"      
## [36] "edit-msvs2008"     "edit-nedit"        "edit-vim-dark"     "edit-vim"          "edit-xcode"       
## [41] "ekvoli"            "fine_blue"         "freya"             "fruit"             "golden"           
## [46] "greenlcd"          "greyscale0"        "greyscale1"        "greyscale2"        "kellys"           
## [51] "leo"               "lucretia"          "manxome"           "maroloccio"        "matrix"           
## [56] "moe"               "molokai"           "moria"             "navajo-night"      "navy"             
## [61] "neon"              "night"             "nightshimmer"      "nuvola"            "olive"            
## [66] "orion"             "oxygenated"        "pablo"             "peaksea"           "print"            
## [71] "rand01"            "rdark"             "relaxedgreen"      "rootwater"         "seashell"         
## [76] "solarized-dark"    "solarized-light"   "tabula"            "tcsoft"            "vampire"          
## [81] "whitengrey"        "xoria256"          "zellner"           "zenburn"           "zmrok"

thm = knit_theme$get("acid")  # parse the theme to a list
knit_theme$set(thm)

opts_knit$set(out.format = NULL)  # restore option

[Package knitr version 1.36 Index]