knitr-package {knitr}	R Documentation

A general-purpose tool for dynamic report generation in R

Description

The knitr package is an implementation of Literate Programming, a programming paradigm that intermingle code chunks (for computing) with prose (for documentation) in the same document.

Details

When the document is compiled, the code chunks can be executed, and the results from computing (text or graphics) are automatically written to the output along with the prose.

This package is an alternative tool to Sweave with a more flexible design and new features like caching and finer control of graphics. It is not limited to LaTeX and is ready to be customized to process other file formats. See the package website in the references for more information and examples.

Note

The pronunciation of knitr is similar to neater or you can think of knitter (but it is single t). The name comes from knit + R (while Sweave = S + weave).

Author(s)

Yihui Xie <https://yihui.org>

References

Full documentation and demos: https://yihui.org/knitr/; FAQ's: https://yihui.org/knitr/faq/

See Also

The core function in this package: knit. If you are an Sweave user, see Sweave2knitr on how to convert Sweave files to knitr.

[Package knitr version 1.39 Index]