plot_crop {knitr}	R Documentation

Crop a plot (remove the edges) using PDFCrop or ImageMagick

Description

The command pdfcrop x x is executed on a PDF plot file, and convert x -trim x is executed for other types of plot files, where x is the plot filename.

Usage


plot_crop(x, quiet = TRUE)

Arguments

x
Filename of the plot.

quiet
Boolean; whether to suppress standard output from the command line utility.

Details

The utility pdfcrop is often shipped with a LaTeX distribution, and convert is a command in ImageMagick (Windows users may have to put the bin path of ImageMagick into the PATH variable).

Value

The original filename.

References

PDFCrop: https://www.ctan.org/pkg/pdfcrop; the convert command in ImageMagick: http://www.imagemagick.org/script/convert.php

[Package knitr version 1.25 Index]