set_alias {knitr}	R Documentation

Set aliases for chunk options

Description

We do not have to use the chunk option names given in knitr; we can set aliases for them. The aliases are a named character vector; the names are aliases and the elements in this vector are the real option names.

Usage

set_alias(...)

Arguments

...
Named arguments. Argument names are aliases, and argument values are real option names.

Value

NULL. opts_knit$get('aliases') is modified as the side effect.

Examples

set_alias(w = "fig.width", h = "fig.height")
# then we can use options w and h in chunk headers instead of fig.width and
# fig.height

[Package knitr version 1.36 Index]