set_header {knitr}	R Documentation

Set the header information

Description

Some output documents may need appropriate header information. For example, for LaTeX output, we need to write \usepackage{tikz} into the preamble if we use tikz graphics; this function sets the header information to be written into the output.

Usage

set_header(...)

Arguments

...
Header components; currently possible components are highlight, tikz and framed, which contain the necessary commands to be used in the HTML header or LaTeX preamble. Note that HTML output does not use the tikz and framed components, since they do not make sense in the context of HTML.

Details

By default, knitr will set up the header automatically. For example, if the tikz device is used, knitr will add \usepackage{tikz} to the LaTeX preamble, and this is done by setting the header component tikz to be a character string: set_header(tikz = '\usepackage{tikz}'). Similary, when we highlight R code using the highlight package (i.e. the chunk option highlight = TRUE), knitr will set the highlight component of the header vector automatically; if the output type is HTML, this component will be different – instead of LaTeX commands, it contains CSS definitions.

For power users, all the components can be modified to adapt to a customized type of output. For instance, we can change highlight to LaTeX definitions of the listings package (and modify the output hooks accordingly), so we can decorate R code using the listings package.

Value

The header vector in opts_knit is set.

Examples

set_header(tikz = "\\usepackage{tikz}")
opts_knit$get("header")

[Package knitr version 1.36 Index]