wrap_rmd {knitr}	R Documentation

Wrap long lines in Rmd files

Description

This function wraps long paragraphs in an R Markdown file. Other elements are not wrapped: the YAML preamble, fenced code blocks, section headers and indented elements. The main reason for wrapping long lines is to make it easier to review differences in version control.

Usage

wrap_rmd(file, width = 80, text = NULL, backup)

Arguments

file
The input Rmd file.

width
The expected line width.

text
A character vector of text lines, as an alternative to file. If text is not NULL, file is ignored.

backup
Path to back up the original file in case anything goes wrong. If set to NULL, no backup is made. The default value is constructed from file by adding __ before the base filename.

Value

If file is provided, it is overwritten; if text is provided, a character vector is returned.

Note

Currently it does not wrap blockquotes or lists (ordered or unordered). This feature may or may not be added in the future.

Examples

wrap_rmd(text = c("```", "1+1", "```", "- a list item", "> a quote", "",
    paste(rep("this is a normal paragraph", 5), collapse = " ")))

[Package knitr version 1.36 Index]